Conflicts:
	code/renderercommon/tr_common.h
	code/renderergl1/tr_backend.c
	code/renderergl2/tr_backend.c

+++ code/renderercommon/tr_common.h
@@ -149,7 +149,7 @@ extern	cvar_t	*r_fontForceAutoHint;
 
 qboolean	R_GetModeInfo( int *width, int *height, float *windowAspect, int mode );
 
-float R_NoiseGet4f( float x, float y, float z, float t );
+float R_NoiseGet4f( float x, float y, float z, double t );
 int R_RandomOn( float t );
 void  R_NoiseInit( void );
